+++ zarrs/tests/array_ops.rs
@@ -66,10 +66,10 @@ fn exercise_array_ops<A: ArrayOps>(array: &A) -> TestResult {
     );
     let _ = array.builder().build_metadata()?;
     assert_eq!(
-        array.subchunk_shape().unwrap(),
-        vec![NonZeroU64::new(1).unwrap(); 2]
+        array.subchunk_grid().unwrap().chunk_shape(&[0, 0])?,
+        Some(vec![NonZeroU64::new(1).unwrap(); 2])
     );
-    assert_eq!(array.subchunk_grid_shape(), &[6, 6]);
+    assert_eq!(array.subchunk_grid().unwrap().grid_shape(), &[6, 6]);
     assert_eq!(array.chunk_origin(&[1, 1])?, [3, 3]);
     assert_eq!(
         array.chunk_shape(&[0, 0])?,